added multiple heads to a group

+++ a2sv_portal_backend/src/app/user-group-season/entities/user-group-season.entity.ts
@@ -18,7 +18,7 @@ export class UserGroupSeason {
   @Field(() => User)
   user: User
 
-  @Field(() => Float)
+  @Field(() => Int)
   rank: number
 
   @Field(() => Int)
